+++ src/cbmc/cbmc_parse_options.cpp
@@ -18,11 +18,12 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <util/string2int.h>
 #include <util/config.h>
-#include <util/language.h>
 #include <util/unicode.h>
 #include <util/memory_info.h>
 #include <util/invariant.h>
 #include <util/exit_codes.h>
+
+#include <langapi/language.h>
 
 #include <ansi-c/c_preprocess.h>
