+++ src/goto-analyzer/goto_analyzer_parse_options.h
@@ -103,7 +103,8 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <util/ui_message.h>
 #include <util/parse_options.h>
-#include <util/language.h>
+
+#include <langapi/language.h>
 
 #include <goto-programs/goto_model.h>
 #include <goto-programs/show_goto_functions.h>
