+++ src/goto-programs/initialize_goto_model.cpp
@@ -14,12 +14,12 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <fstream>
 #include <iostream>
 
-#include <util/language.h>
 #include <util/config.h>
 #include <util/unicode.h>
 
 #include <langapi/mode.h>
 #include <langapi/language_ui.h>
+#include <langapi/language.h>
 
 #include <goto-programs/rebuild_goto_start_function.h>
