+++ src/goto-programs/show_symbol_table.cpp
@@ -14,7 +14,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <iostream>
 #include <memory>
 
-#include <util/language.h>
+#include <langapi/language.h>
 #include <langapi/mode.h>
 
 #include "goto_model.h"
