+++ src/jbmc/jbmc_parse_options.cpp
@@ -18,10 +18,11 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <util/string2int.h>
 #include <util/config.h>
-#include <util/language.h>
 #include <util/unicode.h>
 #include <util/memory_info.h>
 #include <util/invariant.h>
+
+#include <langapi/language.h>
 
 #include <ansi-c/ansi_c_language.h>
