+++ src/langapi/Makefile
@@ -1,5 +1,7 @@
 SRC = language_ui.cpp \
       language_util.cpp \
+      language_file.cpp \
+      language.cpp \
       mode.cpp \
       # Empty last line
 INCLUDES= -I ..
