+++ src/langapi/language_util.cpp
@@ -12,9 +12,9 @@ Author: Daniel Kroening, kroening@cs.cmu.edu
 
 #include <util/symbol_table.h>
 #include <util/namespace.h>
-#include <util/language.h>
 #include <util/std_expr.h>
 
+#include "language.h"
 #include "mode.h"
 
 static std::unique_ptr<languaget> get_language(
